common.rbを削除して星座メソッドはhelperに移動させた
